+++ circus-train-gcp/src/test/java/com/hotels/bdp/circustrain/gcp/context/GCPBeanPostProcessorTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2017 Expedia Inc.
+ * Copyright (C) 2016-2018 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,47 +15,70 @@
  */
 package com.hotels.bdp.circustrain.gcp.context;
 
-import static org.mockito.Matchers.anyString;
-import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyZeroInteractions;
 import static org.mockito.Mockito.when;
 
 import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.fs.FileSystem;
+import org.apache.hadoop.fs.Path;
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Spy;
+import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
 import com.hotels.bdp.circustrain.context.CommonBeans;
+import com.hotels.bdp.circustrain.gcp.BindGoogleHadoopFileSystem;
+import com.hotels.bdp.circustrain.gcp.DistributedFileSystemPathProvider;
+import com.hotels.bdp.circustrain.gcp.FileSystemFactory;
+import com.hotels.bdp.circustrain.gcp.GCPCredentialCopier;
+import com.hotels.bdp.circustrain.gcp.GCPCredentialPathProvider;
 
 @RunWith(MockitoJUnitRunner.class)
 public class GCPBeanPostProcessorTest {
-  private @Spy GCPSecurity security = new GCPSecurity();
-  private @Spy Configuration configuration = new Configuration();
 
-  @Test
-  public void postProcessAfterInitializationWithCorrectBeanName() throws Exception {
-    String beanName = CommonBeans.BEAN_BASE_CONF;
-    GCPBeanPostProcessor processor = new GCPBeanPostProcessor(security);
-    processor.postProcessAfterInitialization(configuration, beanName);
-    verify(security, times(1)).getCredentialProvider();
+  private @Mock GCPCredentialPathProvider credentialPathProvider;
+  private @Mock DistributedFileSystemPathProvider distributedFileSystemPathProvider;
+  private @Mock Configuration configuration;
+  private @Mock BindGoogleHadoopFileSystem bindGoogleHadoopFileSystem;
+  private @Mock FileSystemFactory fileSystemFactory;
+  private @Mock GCPCredentialCopier credentialCopier;
+  private @Mock FileSystem fileSystem;
+
+  private GCPBeanPostProcessor processor;
+
+  @Before
+  public void init() {
+    processor = new GCPBeanPostProcessor(credentialPathProvider, distributedFileSystemPathProvider,
+        bindGoogleHadoopFileSystem, fileSystemFactory, credentialCopier);
   }
 
   @Test
   public void postProcessAfterInitializationWithIncorrectBeanName() throws Exception {
     String beanName = "notBaseConf";
-    GCPBeanPostProcessor processor = new GCPBeanPostProcessor(security);
     processor.postProcessAfterInitialization(configuration, beanName);
-    verify(security, times(0)).getCredentialProvider();
+    verifyZeroInteractions(credentialPathProvider, bindGoogleHadoopFileSystem, credentialCopier);
   }
 
   @Test
-  public void postProcessAfterInitializationWithBlankCredentialProviderDoesntModifyConfiguration() throws Exception {
+  public void postProcessAfterInitializationWithConfigurationBeanProviderPathIsNotNull() throws Exception {
     String beanName = CommonBeans.BEAN_BASE_CONF;
-    GCPBeanPostProcessor processor = new GCPBeanPostProcessor(security);
+    when(credentialPathProvider.newPath()).thenReturn(new Path("/test.json"));
+    when(fileSystemFactory.getFileSystem(configuration)).thenReturn(fileSystem);
+
     processor.postProcessAfterInitialization(configuration, beanName);
-    when(security.getCredentialProvider()).thenReturn("");
-    verify(security, times(1)).getCredentialProvider();
-    verify(configuration, times(0)).set(anyString(), anyString());
+    verify(bindGoogleHadoopFileSystem).bindFileSystem(configuration);
+    verify(credentialCopier)
+        .copyCredentials(fileSystem, configuration, credentialPathProvider, distributedFileSystemPathProvider);
+  }
+
+  @Test
+  public void postProcessAfterInitializationWithConfigurationBeanProviderPathIsNull() throws Exception {
+    String beanName = CommonBeans.BEAN_BASE_CONF;
+    when(credentialPathProvider.newPath()).thenReturn(null);
+
+    processor.postProcessAfterInitialization(configuration, beanName);
+    verifyZeroInteractions(bindGoogleHadoopFileSystem, credentialCopier, fileSystemFactory);
   }
 }